# ARROW-10696: [C++] Add SetBitRunReader