--- a/cpp/src/arrow/util/bit_util_benchmark.cc
+++ b/cpp/src/arrow/util/bit_util_benchmark.cc
@@ -110,6 +110,29 @@ static std::shared_ptr<Buffer> CreateRandomBuffer(int64_t nbytes) {
   return std::move(buffer);
 }
 
+static std::shared_ptr<Buffer> CreateRandomBitsBuffer(int64_t nbits,
+                                                      int64_t set_percentage) {
+  ::arrow::random::RandomArrayGenerator rag(/*seed=*/23);
+  double set_probability =
+      static_cast<double>(set_percentage == -1 ? 0 : set_percentage) / 100.0;
+  std::shared_ptr<Buffer> buffer =
+      rag.Boolean(nbits, set_probability)->data()->buffers[1];
+
+  if (set_percentage == -1) {
+    internal::BitmapWriter writer(buffer->mutable_data(), /*start_offset=*/0,
+                                  /*length=*/nbits);
+    for (int x = 0; x < nbits; x++) {
+      if (x % 2 == 0) {
+        writer.Set();
+      } else {
+        writer.Clear();
+      }
+      writer.Next();
+    }
+  }
+  return buffer;
+}
+
 template <typename DoAnd>
 static void BenchmarkAndImpl(benchmark::State& state, DoAnd&& do_and) {
   int64_t nbytes = state.range(0);
@@ -202,35 +225,37 @@ static void BenchmarkBitmapReader(benchmark::State& state, int64_t nbytes) {
 
 template <typename BitRunReaderType>
 static void BenchmarkBitRunReader(benchmark::State& state, int64_t set_percentage) {
-  ::arrow::random::RandomArrayGenerator rag(/*seed=*/23);
   constexpr int64_t kNumBits = 4096;
-  double set_probability =
-      static_cast<double>(set_percentage == -1 ? 0 : set_percentage) / 100.0;
-  std::shared_ptr<Buffer> buffer =
-      rag.Boolean(kNumBits, set_probability)->data()->buffers[1];
-
-  const uint8_t* bitmap = buffer->data();
-  if (set_percentage == -1) {
-    internal::BitmapWriter writer(buffer->mutable_data(), /*start_offset=*/0,
-                                  /*length=*/kNumBits);
-    for (int x = 0; x < kNumBits; x++) {
-      if (x % 2 == 0) {
-        writer.Set();
-      } else {
-        writer.Clear();
-      }
-      writer.Next();
-    }
-  }
+  auto buffer = CreateRandomBitsBuffer(kNumBits, set_percentage);
 
   for (auto _ : state) {
     {
-      BitRunReaderType reader(bitmap, 0, kNumBits);
+      BitRunReaderType reader(buffer->data(), 0, kNumBits);
       int64_t set_total = 0;
       internal::BitRun br;
       do {
         br = reader.NextRun();
         set_total += br.set ? br.length : 0;
+      } while (br.length != 0);
+      benchmark::DoNotOptimize(set_total);
+    }
+  }
+  state.SetBytesProcessed(state.iterations() * (kNumBits / 8));
+}
+
+template <typename SetBitRunReaderType>
+static void BenchmarkSetBitRunReader(benchmark::State& state, int64_t set_percentage) {
+  constexpr int64_t kNumBits = 4096;
+  auto buffer = CreateRandomBitsBuffer(kNumBits, set_percentage);
+
+  for (auto _ : state) {
+    {
+      SetBitRunReaderType reader(buffer->data(), 0, kNumBits);
+      int64_t set_total = 0;
+      internal::SetBitRun br;
+      do {
+        br = reader.NextRun();
+        set_total += br.length;
       } while (br.length != 0);
       benchmark::DoNotOptimize(set_total);
     }
@@ -346,6 +371,14 @@ static void BitRunReader(benchmark::State& state) {
 
 static void BitRunReaderLinear(benchmark::State& state) {
   BenchmarkBitRunReader<internal::BitRunReaderLinear>(state, state.range(0));
+}
+
+static void SetBitRunReader(benchmark::State& state) {
+  BenchmarkSetBitRunReader<internal::SetBitRunReader>(state, state.range(0));
+}
+
+static void ReverseSetBitRunReader(benchmark::State& state) {
+  BenchmarkSetBitRunReader<internal::ReverseSetBitRunReader>(state, state.range(0));
 }
 
 static void BitmapWriter(benchmark::State& state) {
@@ -477,27 +510,17 @@ static void ReferenceNaiveBitmapReader(benchmark::State& state) {
 BENCHMARK(ReferenceNaiveBitmapReader)->Arg(kBufferSize);
 #endif
 
+void SetBitRunReaderPercentageArg(benchmark::internal::Benchmark* bench) {
+  bench->Arg(-1)->Arg(0)->Arg(10)->Arg(25)->Arg(50)->Arg(60)->Arg(75)->Arg(99);
+}
+
 BENCHMARK(BitmapReader)->Arg(kBufferSize);
 BENCHMARK(BitmapUInt64Reader)->Arg(kBufferSize);
 
-BENCHMARK(BitRunReader)
-    ->Arg(-1)
-    ->Arg(0)
-    ->Arg(10)
-    ->Arg(25)
-    ->Arg(50)
-    ->Arg(60)
-    ->Arg(75)
-    ->Arg(99);
-BENCHMARK(BitRunReaderLinear)
-    ->Arg(-1)
-    ->Arg(0)
-    ->Arg(10)
-    ->Arg(25)
-    ->Arg(50)
-    ->Arg(60)
-    ->Arg(75)
-    ->Arg(99);
+BENCHMARK(BitRunReader)->Apply(SetBitRunReaderPercentageArg);
+BENCHMARK(BitRunReaderLinear)->Apply(SetBitRunReaderPercentageArg);
+BENCHMARK(SetBitRunReader)->Apply(SetBitRunReaderPercentageArg);
+BENCHMARK(ReverseSetBitRunReader)->Apply(SetBitRunReaderPercentageArg);
 
 BENCHMARK(VisitBits)->Arg(kBufferSize);
 BENCHMARK(VisitBitsUnrolled)->Arg(kBufferSize);