… cpp/src/arrow/util/spaced.h
… cpp/src/parquet/encoding.cc
… cpp/src/parquet/statistics.cc