A specialized bitmap reader that yields runs of set bits, for use cases where reset bits (e.g. null bits) don't need any handling.

On some use cases it can be significantly faster than the alternatives.