@pitrou 2020-11-25
Also some speedups on no-null filters:
* before:
```
FilterInt64FilterNoNulls/524288/0          19993 ns        19988 ns        34897 bytes_per_second=24.4281G/s data null%=0 mask null%=0 select%=99.9 size=524.288k
FilterInt64FilterNoNulls/524288/1         249210 ns       249153 ns         2805 bytes_per_second=1.95976G/s data null%=0 mask null%=0 select%=50 size=524.288k
FilterInt64FilterNoNulls/524288/2          31896 ns        31889 ns        21867 bytes_per_second=15.3118G/s data null%=0 mask null%=0 select%=1 size=524.288k

FilterStringFilterNoNulls/524288/0         54006 ns        53996 ns        12804 bytes_per_second=9.04295G/s data null%=0 mask null%=0 select%=99.9 size=524.288k
FilterStringFilterNoNulls/524288/1        256494 ns       256438 ns         2721 bytes_per_second=1.90409G/s data null%=0 mask null%=0 select%=50 size=524.288k
FilterStringFilterNoNulls/524288/2         16881 ns        16878 ns        40537 bytes_per_second=28.9302G/s data null%=0 mask null%=0 select%=1 size=524.288k
```

* after:
```
FilterInt64FilterNoNulls/524288/0          14416 ns        14413 ns        49758 bytes_per_second=33.8771G/s data null%=0 mask null%=0 select%=99.9 size=524.288k
FilterInt64FilterNoNulls/524288/1         151899 ns       151870 ns         4514 bytes_per_second=3.21512G/s data null%=0 mask null%=0 select%=50 size=524.288k
FilterInt64FilterNoNulls/524288/2           8578 ns         8577 ns        82823 bytes_per_second=56.931G/s data null%=0 mask null%=0 select%=1 size=524.288k

FilterStringFilterNoNulls/524288/0         45116 ns        45104 ns        15887 bytes_per_second=10.8256G/s data null%=0 mask null%=0 select%=99.9 size=524.288k
FilterStringFilterNoNulls/524288/1        146822 ns       146796 ns         4746 bytes_per_second=3.32625G/s data null%=0 mask null%=0 select%=50 size=524.288k
FilterStringFilterNoNulls/524288/2          5438 ns         5437 ns       126135 bytes_per_second=89.8084G/s data null%=0 mask null%=0 select%=1 size=524.288k
```

(couldn't run a diff because of ARROW-10738)